app/src/main/java/com/dicoding/prayogo/footballLeagueApp/presenter/MainPresenter.kt
package com.dicoding.prayogo.footballLeagueApp.presenter

import com.dicoding.prayogo.footballLeagueApp.api.ApiRepository
import com.dicoding.prayogo.footballLeagueApp.api.TheSportDBApi
import com.dicoding.prayogo.footballLeagueApp.model.LeagueResponse
import com.dicoding.prayogo.footballLeagueApp.view.MainView
import com.google.gson.Gson
import kotlinx.coroutines.GlobalScope
import kotlinx.coroutines.launch
import com.dicoding.prayogo.footballLeagueApp.util.CoroutineContextProvider


class MainPresenter(
    private val view: MainView,
    private val apiRepository: ApiRepository,
    private val gson: Gson,
    private val context: CoroutineContextProvider = CoroutineContextProvider()
) {
    fun getLeagueList(country: String?) {
        view.showLoading()
        GlobalScope.launch(context.main) {
            val data = gson.fromJson(
                apiRepository
                    .doRequest(TheSportDBApi.getLeague(country)).await(),
                LeagueResponse::class.java
            )
            view.hideLoading()
            view.showLeagueList(data.countrys)
        }
    }
}
